Implemented ignoring back-as-home in windows phones

+++ www/js/app.js
@@ -66,9 +66,13 @@ angular.module('relatedwords', ['ionic', 'relatedwords.controllers', 'relatedwor
   $ionicPlatform.registerBackButtonAction(function (event){
     var currentState = $ionicHistory.currentStateName();
     if(currentState == "collections"){
-      // navigator.app.exitApp();//exit the app
-      backAsHome.trigger();//exitApp as home button (don't kill the app)
-      // event.preventDefault();//don't do anything
+      if(isEdge){
+        navigator.app.exitApp();//exit the app
+      }
+      else{
+        backAsHome.trigger();//exitApp as home button (don't kill the app)
+      }
+        //event.preventDefault();//don't do anything
     }
     else{
       //this can also be accompolished using $ionicPlatform.onHardwareBackButton
